+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- **Breaking**: module: GetParamErrorとScriptModuleCallHandleErrorを統合する
 - filter: FilterAudioProcのレシーバを`&self`にする
 
 ### デモプラグイン